Moved Transport.cs and ITransport.cs into Core/*
